separate app from app stack
